Change hash_internal to return u64 (#758)
Commit ef0d257501 in erlang/otp changed enif_hash to use
erts_internal_salted_hash, which returns a unsigned 64-bit integer, as well as
accepting a 64-bit salt. This commit was released as part of OTP 27.

+++ rustler_tests/native/rustler_test/src/test_term.rs
@@ -45,7 +45,7 @@ pub fn term_cmp<'a>(a: Term<'a>, b: Term<'a>) -> Atom {
 }
 
 #[rustler::nif]
-pub fn term_internal_hash(term: Term, salt: u32) -> u32 {
+pub fn term_internal_hash(term: Term, salt: u64) -> u64 {
     term.hash_internal(salt)
 }
